Minor speedup
Changed it such that instead of copying A, B into two distinct arrays,
it just copies into 1 giant array. (Main change was in bijectB with the
addition of the n2 term in the Bt lookup).

Also, loop ordering was adjusted in the bijectA function.

Will attempt blocking tonight (but don't see why it'll increase the
speed as ~1500 length array will fill the L1 cache)