Added parallel qsort.
I just modified the code Marshall had referred to. The code divides the array into as many chunks as the number of threads and uses qsort to sort the chunks in parallel. Then it sorts the array consisting of the sorted chunks.

It turned out dividing the array into sorted chunks is enough to boost the performance. The second step, that is sorting the entire array, only adds to the computations and does not improve the locality that much. That is why I have commented it. You can test it by uncommenting lines 292-294 in sph.c