Implemented the EG algorithm. The topics.txt is quite different from that of the active set algorithm. This could happen according to Bindel's comments on piazza, but I do not know how much difference is acceptable. You guys can get started with the parallelization until I get answers to my questions on piazza.
Bindel recommended line 164 in anchor_topics, but the code runs faster when I comment it. I am not sure how important this is since there is no clear way of testing the code.

BTW, bindel has said on piazza that we can stick to the active set algorithm as even his implementation of the EG algorithm turned out to be slower.